fix(admin): run transcript groups in one workflow step
